+++ static/js/nzbhydra.js
@@ -2304,7 +2304,6 @@ function SearchHistoryController($scope, $state, StatsService, history) {
     $scope.searchRequests = history.data.searchRequests;
     $scope.totalRequests = history.data.totalRequests;
 
-
     $scope.pageChanged = function (newPage) {
         getSearchRequestsPage(newPage);
     };
@@ -2373,6 +2372,10 @@ function SearchHistoryController($scope, $state, StatsService, history) {
         }
         if (request.tvtitle != null) {
             return request.tvtitle;
+        }
+
+        if (!request.query && !request.identifier_key && !request.season && !request.episode) {
+            return "Update query";
         }
         return request.query;
     }
@@ -4816,12 +4819,8 @@ function getIndexerPresets() {
             host: "https://6box.me"
         },
         {
-            name: "6box nzedb",
-            host: "https://nzedb.6box.me"
-        },
-        {
-            name: "6box nntmux",
-            host: "https://nn-tmux.6box.me"
+            name: "6box sptweb",
+            host: "https://6box.me/spotweb"
         },
         {
             name: "DogNZB",