+++ ui-src/js/config-fields-service.js
@@ -898,12 +898,8 @@ function getIndexerPresets() {
             host: "https://6box.me"
         },
         {
-            name: "6box nzedb",
-            host: "https://nzedb.6box.me"
-        },
-        {
-            name: "6box nntmux",
-            host: "https://nn-tmux.6box.me"
+            name: "6box sptweb",
+            host: "https://6box.me/spotweb"
         },
         {
             name: "DogNZB",